fix config for zustand
